Added API functionality to adding search player
When a user clicks the add player on a searched player, the API now
successfully adds that player to their list and updates the list
accordingly. Separate methods were written for the adding of a player
from the search results as these "add player" buttons were added to the
DOM dynamically and required different handling.